rename log handler
