+++ exporters/otlp/src/otlp_http_exporter.cc
@@ -16,7 +16,7 @@
 #include "opentelemetry/proto/collector/trace/v1/trace_service.pb.h"
 
 #include "opentelemetry/exporters/otlp/protobuf_include_suffix.h"
-#include "opentelemetry/sdk/common/global_error_handler.h"
+#include "opentelemetry/sdk_config.h"
 
 #include <condition_variable>
 #include <fstream>
