+++ sdk/include/opentelemetry/sdk/common/global_log_handler.h
